requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
boto3>=1.34.0
pydantic>=2.5.0
requests>=2.31.0
pytest>=7.4.0
